src/agents/skill/requirements.txt
python-dotenv
azure-identity
azure-ai-projects==2.0.0b3
openai
mcp
